refactor: split pricing table
